+++ src/render/shaders.js
@@ -63,10 +63,11 @@ export const meshVertexShader = `
 
 export const meshFragmentShader = `
     precision mediump float;
+    uniform float alpha;
     varying vec3 fragColor;
 
     void main() {
-        gl_FragColor = vec4(fragColor, 1.0);
+        gl_FragColor = vec4(fragColor, alpha);
     }
 `;